Model Creation I
Impemented modelWithNTopics() method, which creates models for each topic quantity specified in an integer vector argument.

This method will be used in a nested loop structure for a black box method for finding an optimal model.  Additionally, it may be used as a standalone method to create models.

TO-DO:  Create progress indicator for numModelsCreated/TotalModelsToCreate which functions if this method is used standalone and integrates seamlessly with an outer loop structure if used in the black box container.