Prediction
Implemented the Model.predict() instance method, which permits users to easily make predictions from a List of Strings using a Model object.

Additionally, implemented predictFromFile(), which provides the same functionality except it permits direct use of a Model saved to file instead of requiring a Model object in the namespace.